Added functionality to add variables and methods to the boxes as well as updated the boxes on screen.
